fix bugs commit

+++ server/src/routes/memories.ts
@@ -1,16 +1,25 @@
 import { FastifyInstance } from 'fastify'
 import { prisma } from '../lib/prisma'
-import { z } from 'zod'
-// import { request } from 'http'
-// import { request } from 'http'
+import z from 'zod'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify() // se o token não for válido, vai bloquear e não descer para as linhas de baixo
+  })
+
+  app.get('/memories', async (request) => {
+    // trazer só as memórias do usuário logado
+
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
     })
+
+    // esse map não deixa retornar todo o texto de content, pois só precisa de algumas poucas palavras ali no feed do site
     return memories.map((memory) => {
       return {
         id: memory.id,
@@ -20,23 +29,31 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (request) => {
-    // const { id } = request.params
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
+    // ta vendo se esse id segue a estrutura que eu defini no zod
     const { id } = paramsSchema.parse(request.params)
 
+    // encontre a memória com esse id, se não disparar um erro
     const memory = await prisma.memory.findUniqueOrThrow({
       where: {
         id,
       },
     })
+
+    // checando se a memória não for publica e o id do usuário que escrever a memória for diferente do do id da requisição, vai retornar o erro 401 de não encontrado
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
     return memory
   })
 
-  app.post('/memories', async (request) => {
+  // o boolean do zod é 1,
+  app.post('/memories/:id', async (request) => {
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
@@ -50,13 +67,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: 'd3ef067c-6198-4b91-833c-6fdcef63e3c6',
+        userId: request.user.sub,
       },
     })
+
     return memory
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -71,7 +89,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -81,21 +109,31 @@ export async function memoriesRoutes(app: FastifyInstance) {
         isPublic,
       },
     })
+
     return memory
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params)
 
-    const memory = await prisma.memory.delete({
+    const memory = await prisma.memory.findUniqueOrThrow({
       where: {
         id,
       },
     })
-    return memory
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    await prisma.memory.delete({
+      where: {
+        id,
+      },
+    })
   })
 }
